24 hour high and low temperatures

+++ usb_weather_reading.h
@@ -23,7 +23,7 @@ public:
 		values it wrote as the logger data.
 		(see:http://sandaysoft.com/forum/viewtopic.php?f=13&t=10524)
 	*/
-	uint8_t  delay;						// minutes since last recording
+	uint32_t  delay;					// minutes since last recording
 	double indoor_humidity;				// percent
 	double indoor_temperature;			// degrees C
 	double outdoor_humidity;			// percent
